bug fix, now is workable

+++ step_tear/fg/serial/fgNoMbb.py
@@ -18,8 +18,8 @@ def main():
     min_y = float(sys.argv[2])
     max_x = float(sys.argv[3])
     max_y = float(sys.argv[4])
-    bucket_size = int(sys.argv[5])
-    num_object = int(sys.argv[6])
+    bucket_size = float(sys.argv[5])
+    num_object = float(sys.argv[6])
 
     span_x = max_x - min_x
     span_y = max_y - min_y
@@ -28,9 +28,12 @@ def main():
     x_split = y_split = 1
     if span_y > span_x:
         # We prefer to split into more-square regions than long rectangles
-        y_split = math.ceil(math.sqrt(num_object / bucket_size * (span_y /
+# 2267313 102137873 130890.0 130146.0
+        #print num_object, bucket_size, span_y, span_x
+        y_split = math.ceil(math.sqrt(float(num_object) / bucket_size * (span_y /
                                                                   span_x)))
-        x_split = math.ceil(num_object / bucket_size / y_split)
+
+        x_split = math.ceil(float(num_object) / bucket_size / y_split)
     else:
         x_split = math.ceil(math.sqrt(num_object / bucket_size * (span_x /
                                                                   span_y)))
@@ -39,13 +42,13 @@ def main():
     #sys.stderr.write("num_x_split", str(x_split))
     #sys.stderr.write("num_y_split", str(y_split))
 
-    xtile = 1.0 / x_split
-    ytile = 1.0 / y_split
+    xtile = 1.0 / float(x_split)
+    ytile = 1.0 / float(y_split)
     id = 0
     x = 0
     y = 0
-    for  x in range(0, x_split) :
-        for y in range(0, y_split) :
+    for  x in range(0, int(x_split)) :
+        for y in range(0, int(y_split)) :
             id += 1
     #        print("\t".join((str(id), str(min_x + x * xtile ),
     #                       str(min_y + y * ytile),
